Move the "sweep" process to Sweep::main()
